feat(config): change config file to accept only functions, split as v1, and use symbol in type for future compat
