+++ libs/notios-config/src/define_notios_config.ts
@@ -1,7 +1,9 @@
 import { NotiosConfig } from './interfaces/notios-config';
 
-const defineNotiosConfig = (config: NotiosConfig): NotiosConfig => {
-  return config;
+export type ConfigFn = (defaultConfig: NotiosConfig) => NotiosConfig;
+
+const defineNotiosConfig = (configFn: ConfigFn): ConfigFn => {
+  return configFn;
 };
 
 export default defineNotiosConfig;
