+++ libs/notios-config/src/index.ts
@@ -1,2 +1,1 @@
-export { default as defaultConfig } from './default_config';
 export { default as defineNotiosConfig } from './define_notios_config';
